@@ -4,7 +4,8 @@ Type annotations for botocore.docs.shape module.
 Copyright 2025 Vlad Emelianov
 """
 
-from typing import Any, Mapping
+from collections.abc import Mapping
+from typing import Any
 
 from botocore.docs.bcdoc.restdoc import DocumentStructure
 from botocore.hooks import BaseEventHooks
